+++ src/main/java/lotto/Application.java
@@ -1,8 +1,12 @@
 package lotto;
 
+import camp.nextstep.edu.missionutils.Console;
+
+import static camp.nextstep.edu.missionutils.Console.readLine;
+import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
+
 public class Application {
     public static void main(String[] args) {
-
 
     }
 }
