+++ src/main/java/lotto/utils/Validation.java
@@ -1,0 +1,23 @@
+package lotto.utils;
+
+import java.util.List;
+
+public class Validation {
+
+    public static void validate(List<Integer> numbers) {
+        validateSize(numbers);
+        validateDuplication(numbers);
+    }
+
+    private static void validateDuplication(List<Integer> numbers) {
+        if (numbers.stream().distinct().count() != 6) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    private static void validateSize(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
